fix(general): Silence narrowing and signedness conversion warnings
ROOT's Minimizer API takes unsigned int variable indices while every caller
here counts with std::size_t, so SetVariable/FixVariable/IsFixedVariable and
the ParameterWrapper subscript now cast explicitly at the call site rather
than relying on an implicit narrowing conversion. Same for
get_general_parameter_name/get_detector_parameter_name in Parameter.h and the
system_clock seed in DataBase.cpp.

ReactorData's entry loop counted with unsigned int against a size_t size, and
Oscillator's index loop did the same; both now use size_t throughout. Also
adds the missing trailing newline to ReactorData.cpp and ParameterWrapper.cpp.

No behavioural change -- every value involved was already in range.

+++ libraries/io/DoubleChooz/DataBase.cpp
@@ -622,7 +622,7 @@ namespace io::dc {
 
     construct_correlation_matrices();
 
-    std::default_random_engine gen(std::chrono::system_clock::now().time_since_epoch().count());
+    std::default_random_engine gen(static_cast<unsigned int>(std::chrono::system_clock::now().time_since_epoch().count()));
 
     try {
       for (auto detector : {ND, FDI, FDII}) {
